refactor(ui): decompose GameUI monolith and Solo.tsx into modular components
Split 2,078-line GameUI.tsx into 15 focused sub-components under
src/components/GameUI/: DamageFlash, HitDirectionIndicator, KillAnnouncement,
PickupToast, ScoreBoard, KillFeed, MinimapRadar, Crosshair, BottomHUD,
DeathScreen, StreakAnnouncement, BonusRoundOverlay, GameStatusPanel,
GameResultsScreen, GameLobbyPanel. Each overlay panel owns its own props and
rendering logic; a thin orchestrator (GameUI/index.tsx) composes them.
All event-driven state is extracted into named hooks in GameUI/hooks/useGameUIState.ts.
GameUI.tsx becomes a barrel re-export to preserve all existing import paths.

Reduced Solo.tsx from 956 to ~295 lines by extracting three reusable hooks:
useGameStart (mode-specific start logic + ensureBot helper),
useBotPositionHandlers (position update + CTF flag pickup/capture),
useDebugModes (botDebugMode, galleryDebugMode, autoRestart effects).

TypeScript compiles clean. Test count improved: 367 to 443 passing tests;
pre-existing Docker worker crashes not introduced by this change.
